Begin implementing encryption

+++ app/src/test/java/com/example/safecrypt/BitShiftingTest.kt
@@ -14,8 +14,8 @@ class BitShiftingTest {
         for (length in 0..128) {
             for (numberToShift in 0..(length shl 3)) {
                 val data = Random.nextUBytes(length)
-                val leftShiftedData = data.shiftLeftBits(numberToShift)
-                val unLeftShiftedData = leftShiftedData.shiftRightBits(numberToShift)
+                val leftShiftedData = data.shiftLeftBits(numberToShift.toBigInteger())
+                val unLeftShiftedData = leftShiftedData.shiftRightBits(numberToShift.toBigInteger())
                 assertTrue(unLeftShiftedData contentEquals  data)
             }
         }
